add autorization user

+++ model/index.js
@@ -1,36 +1,35 @@
 const Contact = require("./schemas/contact");
 
-
-
-const getAll = async () => {
-  const results = await Contact.find({})
-  return results
-}
-
-const getContactById = async (id) => {
-  const result = await Contact.findOne({ _id: id })
-  return result
-}
-
-const removeContact = async (id) => {
-  const result = await Contact.findByIdAndRemove({ _id: id })
-  return result
+const getAll = async (userId) => {
+  const results = await Contact.find({ owner: userId });
+  return results;
 };
 
-const addContact = async (body) => {
-  const result = await Contact.create(body)
-  return result
+const getContactById = async (userId, id) => {
+  const result = await Contact.findOne({ _id: id, owner: userId });
+  return result;
 };
 
-const updateContact = async (id, body) => {
+const removeContact = async (userId, id) => {
+  const result = await Contact.findByIdAndRemove({ _id: id, owner: userId });
+  return result;
+};
+
+const addContact = async (userId, body) => {
+  const result = await Contact.create(body);
+  return result;
+};
+
+const updateContact = async (userId, id, body) => {
   const result = await Contact.findOneAndUpdate(
     {
       _id: id,
+      owner: userId,
     },
     { ...body },
-    { new: true },
-  )
-  return result
+    { new: true }
+  );
+  return result;
 };
 
 module.exports = {
